+++ ChangeLog.md
@@ -4,7 +4,9 @@
 - Obfuscation supports removing all captions
 - Properties dialog for graph properties added
 - "FastRenderingNodeCountLimit" optimization removed, seems no longer be needed
-- Package spec: wildcard handling improved
+- Package spec
+  - wildcard handling improved
+  - patterns without wildcards changed to "equals" match
 
 ## 8.2 - 2025-02-26